+++ components/brave_news/browser/feed_sampling.cc
@@ -5,114 +5,20 @@
 
 #include "brave/components/brave_news/browser/feed_sampling.h"
 
-#include <algorithm>
 #include <numeric>
 #include <optional>
 #include <vector>
 
 #include "base/containers/contains.h"
-#include "base/containers/flat_set.h"
 #include "base/functional/bind.h"
 #include "base/rand_util.h"
-#include "brave/components/brave_news/browser/channels_controller.h"
-#include "brave/components/brave_news/browser/publishers_controller.h"
-#include "brave/components/brave_news/browser/signal_calculator.h"
-#include "brave/components/brave_news/common/brave_news.mojom-forward.h"
-#include "brave/components/brave_news/common/brave_news.mojom.h"
-#include "brave/components/brave_news/common/features.h"
 
 namespace brave_news {
 
-namespace {
-
-double GetSubscribedWeight(const mojom::FeedItemMetadataPtr& article,
-                           const std::vector<mojom::Signal*>& signals) {
-  return std::reduce(
-      signals.begin(), signals.end(), 0.0, [](double prev, const auto* signal) {
-        return prev + signal->subscribed_weight / signal->article_count;
-      });
-}
-
-double GetPopRecency(const mojom::FeedItemMetadataPtr& article) {
-  const double pop_recency_half_life_in_hours =
-      features::kBraveNewsPopScoreHalfLife.Get();
-
-  auto& publish_time = article->publish_time;
-
-  double popularity = std::min(article->pop_score, 100.0) / 100.0 +
-                      features::kBraveNewsPopScoreMin.Get();
-  double multiplier = publish_time > base::Time::Now() - base::Hours(5) ? 2 : 1;
-  auto dt = base::Time::Now() - publish_time;
-
-  return multiplier * popularity *
-         pow(0.5, dt.InHours() / pop_recency_half_life_in_hours);
-}
-
-// Gets all relevant signals for an article.
-// **Note:** Importantly, this function returns the Signal from the publisher
-// first, and |GetArticleWeight| depends on this to determine whether the
-// Publisher has been visited.
-std::vector<mojom::Signal*> GetSignals(
-    const std::string& locale,
-    const mojom::FeedItemMetadataPtr& article,
-    const Publishers& publishers,
-    const Signals& signals) {
-  std::vector<mojom::Signal*> result;
-  auto it = signals.find(article->publisher_id);
-  if (it != signals.end()) {
-    result.push_back(it->second.get());
-  }
-
-  auto publisher_it = publishers.find(article->publisher_id);
-  if (publisher_it == publishers.end()) {
-    return result;
-  }
-
-  for (const auto& locale_info : publisher_it->second->locales) {
-    if (locale_info->locale != locale) {
-      continue;
-    }
-    for (const auto& channel : locale_info->channels) {
-      auto signal_it = signals.find(channel);
-      if (signal_it == signals.end()) {
-        continue;
-      }
-      result.push_back(signal_it->second.get());
-    }
-  }
-  return result;
-}
-
-ArticleMetaData GetArticleWeight(const mojom::FeedItemMetadataPtr& article,
-                                 const std::vector<mojom::Signal*>& signals,
-                                 std::vector<std::string> publisher_channels,
-                                 const bool& discoverable) {
-  // We should have at least one |Signal| from the |Publisher| for this source.
-  CHECK(!signals.empty());
-
-  const auto subscribed_weight = GetSubscribedWeight(article, signals);
-  const double source_visits_min = features::kBraveNewsSourceVisitsMin.Get();
-  const double source_visits_projected =
-      source_visits_min + signals.at(0)->visit_weight * (1 - source_visits_min);
-  const auto pop_recency = GetPopRecency(article);
-
-  ArticleMetaData weight;
-  weight.pop_recency = pop_recency;
-  weight.weighting =
-      (source_visits_projected + subscribed_weight) * pop_recency;
-  weight.visited = signals.at(0)->visit_weight != 0;
-  weight.subscribed = subscribed_weight != 0,
-  weight.discoverable = discoverable;
-  weight.channels = base::flat_set<std::string>(std::move(publisher_channels));
-  return weight;
-}
-
-}  // namespace
-
-ArticleMetaData::ArticleMetaData() = default;
-ArticleMetaData::~ArticleMetaData() = default;
-ArticleMetaData::ArticleMetaData(ArticleMetaData&&) = default;
-ArticleMetaData& ArticleMetaData::operator=(ArticleMetaData&&) = default;
+ArticleMetadata::ArticleMetadata() = default;
+ArticleMetadata::~ArticleMetadata() = default;
+ArticleMetadata::ArticleMetadata(ArticleMetadata&&) = default;
+ArticleMetadata& ArticleMetadata::operator=(ArticleMetadata&&) = default;
 
 ContentGroup SampleContentGroup(
     const std::vector<ContentGroup>& eligible_content_groups) {
@@ -190,7 +96,7 @@ std::optional<size_t> PickRouletteWithWeighting(const ArticleInfos& articles,
 std::optional<size_t> PickRoulette(const ArticleInfos& articles) {
   return PickRouletteWithWeighting(
       articles, base::BindRepeating([](const mojom::FeedItemMetadataPtr& data,
-                                       const ArticleMetaData& meta) {
+                                       const ArticleMetadata& meta) {
         return meta.subscribed ? meta.weighting : 0;
       }));
 }
@@ -201,73 +107,12 @@ std::optional<size_t> PickChannelRoulette(const std::string& channel,
       articles, base::BindRepeating(
                     [](const std::string& channel,
                        const mojom::FeedItemMetadataPtr& metadata,
-                       const ArticleMetaData& weight) {
+                       const ArticleMetadata& weight) {
                       return base::Contains(weight.channels, channel)
                                  ? weight.weighting
                                  : 0.0;
                     },
                     channel));
-}
-
-ArticleInfos GetArticleInfos(const std::string& locale,
-                             const FeedItems& feed_items,
-                             const Publishers& publishers,
-                             const Signals& signals) {
-  ArticleInfos articles;
-  base::flat_set<GURL> seen_articles;
-  base::flat_set<std::string> non_discoverable_publishers;
-
-  for (const auto& [publisher_id, publisher] : publishers) {
-    auto channels = GetChannelsForPublisher(locale, publisher);
-    if (base::ranges::any_of(kSensitiveChannels,
-                             [&](const std::string& channel) {
-                               return base::Contains(channels, channel);
-                             })) {
-      non_discoverable_publishers.insert(publisher_id);
-    }
-  }
-
-  for (const auto& item : feed_items) {
-    if (item.is_null()) {
-      continue;
-    }
-    if (item->is_article()) {
-      // Because we download feeds from multiple locales, it's possible there
-      // will be duplicate articles, which we should filter out.
-      if (seen_articles.contains(item->get_article()->data->url)) {
-        continue;
-      }
-      auto& article = item->get_article();
-
-      seen_articles.insert(article->data->url);
-
-      auto article_signals =
-          GetSignals(locale, article->data, publishers, signals);
-
-      // If we don't have any signals for this article, or the source this
-      // article comes from has been disabled then filter it out.
-      if (article_signals.empty() ||
-          base::ranges::any_of(article_signals, [](const auto* signal) {
-            return signal->disabled;
-          })) {
-        continue;
-      }
-
-      const bool discoverable = !base::Contains(non_discoverable_publishers,
-                                                article->data->publisher_id);
-
-      auto channels = GetChannelsForPublisher(
-          locale, publishers.at(article->data->publisher_id));
-      ArticleInfo pair =
-          std::tuple(article->data->Clone(),
-                     GetArticleWeight(article->data, article_signals,
-                                      std::move(channels), discoverable));
-
-      articles.push_back(std::move(pair));
-    }
-  }
-
-  return articles;
 }
 
 }  // namespace brave_news
