+++ components/brave_news/browser/feed_sampling.h
@@ -15,14 +15,11 @@
 
 #include "base/functional/callback_forward.h"
 #include "base/rand_util.h"
-#include "brave/components/brave_news/browser/publishers_controller.h"
-#include "brave/components/brave_news/browser/signal_calculator.h"
 #include "brave/components/brave_news/common/brave_news.mojom-forward.h"
 
 namespace brave_news {
 
-// An ArticleWeight has a few different components
-struct ArticleMetaData {
+struct ArticleMetadata {
   // The pop_recency of the article. This is used for discover cards, where we
   // don't consider the subscription status or visit_weighting.
   double pop_recency = 0;
@@ -47,22 +44,22 @@ struct ArticleMetaData {
   // All the channels this Article belongs to.
   base::flat_set<std::string> channels;
 
-  ArticleMetaData();
-  ArticleMetaData(const ArticleMetaData&) = delete;
-  ArticleMetaData& operator=(const ArticleMetaData&) = delete;
-  ArticleMetaData& operator=(ArticleMetaData&&);
-  ArticleMetaData(ArticleMetaData&&);
-  ~ArticleMetaData();
+  ArticleMetadata();
+  ArticleMetadata(const ArticleMetadata&) = delete;
+  ArticleMetadata& operator=(const ArticleMetadata&) = delete;
+  ArticleMetadata& operator=(ArticleMetadata&&);
+  ArticleMetadata(ArticleMetadata&&);
+  ~ArticleMetadata();
 };
 
-using ArticleInfo = std::tuple<mojom::FeedItemMetadataPtr, ArticleMetaData>;
+using ArticleInfo = std::tuple<mojom::FeedItemMetadataPtr, ArticleMetadata>;
 using ArticleInfos = std::vector<ArticleInfo>;
 
 // Gets a weighting for a specific article. This determines how likely an
 // article is to be chosen.
 using GetWeighting =
     base::RepeatingCallback<double(const mojom::FeedItemMetadataPtr& data,
-                                   const ArticleMetaData& meta)>;
+                                   const ArticleMetadata& meta)>;
 
 // PickArticles is a strategy used to pick articles (for example, taking the
 // first article). Different feeds use different strategies for picking
@@ -108,11 +105,6 @@ std::optional<size_t> PickRouletteWithWeighting(const ArticleInfos& articles,
 std::optional<size_t> PickRoulette(const ArticleInfos& articles);
 std::optional<size_t> PickChannelRoulette(const std::string& channel,
                                           const ArticleInfos& articles);
-
-ArticleInfos GetArticleInfos(const std::string& locale,
-                             const FeedItems& feed_items,
-                             const Publishers& publishers,
-                             const Signals& signals);
 
 }  // namespace brave_news
